Change exec default to env vars, added `--server`
