+++ exec.go
@@ -1,11 +1,7 @@
 package main
 
 import (
-	"encoding/json"
 	"io/ioutil"
-	"log"
-	"net"
-	"net/http"
 	"os"
 	"os/exec"
 	"strings"
@@ -17,14 +13,14 @@ import (
 )
 
 type ExecCommandInput struct {
-	Profile  string
-	Command  string
-	Args     []string
-	Keyring  keyring.Keyring
-	Duration time.Duration
-	MfaToken string
-	WriteEnv bool
-	Signals  chan os.Signal
+	Profile     string
+	Command     string
+	Args        []string
+	Keyring     keyring.Keyring
+	Duration    time.Duration
+	MfaToken    string
+	StartServer bool
+	Signals     chan os.Signal
 }
 
 func ExecCommand(ui Ui, input ExecCommandInput) {
@@ -69,14 +65,17 @@ func ExecCommand(ui Ui, input ExecCommandInput) {
 		env = overwriteEnv(env, "AWS_REGION", region)
 	}
 
-	if err := startCredentialsServer(ui, creds); err != nil {
-		ui.Debug.Println("Failed to start local credentials server", err)
-		input.WriteEnv = true
-	} else {
-		ui.Debug.Println("Listening on local credentials server")
+	writeEnv := true
+
+	if input.StartServer {
+		if err := startCredentialsServer(ui, creds); err != nil {
+			ui.Error.Fatal(err)
+		} else {
+			writeEnv = false
+		}
 	}
 
-	if input.WriteEnv {
+	if writeEnv {
 		ui.Debug.Println("Writing temporary credentials to ENV")
 
 		env = overwriteEnv(env, "AWS_ACCESS_KEY_ID", val.AccessKeyID)
@@ -111,50 +110,6 @@ func ExecCommand(ui Ui, input ExecCommandInput) {
 			os.Exit(waitStatus.ExitStatus())
 		}
 	}
-}
-
-func checkServerRunning(bind string) bool {
-	_, err := net.DialTimeout("tcp", bind, time.Millisecond*10)
-	return err == nil
-}
-
-func startCredentialsServer(ui Ui, creds *VaultCredentials) error {
-	if !checkServerRunning(metadataBind) {
-		ui.Error.Println("Starting `aws-vault server` as root in the background")
-		cmd := exec.Command("sudo", "-b", os.Args[0], "server")
-		cmd.Stdin = os.Stdin
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			return err
-		}
-	}
-
-	l, err := net.Listen("tcp", localServerBind)
-	if err != nil {
-		return err
-	}
-
-	log.Printf("Local instance role server running on %s", l.Addr())
-	go http.Serve(l, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		val, err := creds.Get()
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusGatewayTimeout)
-			return
-		}
-
-		json.NewEncoder(w).Encode(map[string]interface{}{
-			"Code":            "Success",
-			"LastUpdated":     time.Now().Format(awsTimeFormat),
-			"Type":            "AWS-HMAC",
-			"AccessKeyId":     val.AccessKeyID,
-			"SecretAccessKey": val.SecretAccessKey,
-			"Token":           val.SessionToken,
-			"Expiration":      creds.Expires().Format(awsTimeFormat),
-		})
-	}))
-
-	return nil
 }
 
 // write out a config excluding role switching keys
