+++ main.go
@@ -38,6 +38,7 @@ func main() {
 		execProfile      = exec.Arg("profile", "Name of the profile").Required().String()
 		execSessDuration = exec.Flag("session-ttl", "Expiration time for aws session").Default("4h").OverrideDefaultFromEnvar("AWS_SESSION_TTL").Short('t').Duration()
 		execMfaToken     = exec.Flag("mfa-token", "The mfa token to use").Short('m').String()
+		execServer       = exec.Flag("server", "Run the server in the background for credentials").Short('s').Bool()
 		execCmd          = exec.Arg("cmd", "Command to execute").Default(os.Getenv("SHELL")).String()
 		execCmdArgs      = exec.Arg("args", "Command arguments").Strings()
 		rm               = kingpin.Command("rm", "Removes credentials")
@@ -98,13 +99,14 @@ func main() {
 		signal.Notify(signals, os.Interrupt, os.Kill)
 
 		ExecCommand(ui, ExecCommandInput{
-			Profile:  *execProfile,
-			Command:  *execCmd,
-			Args:     *execCmdArgs,
-			Keyring:  keyring,
-			Duration: *execSessDuration,
-			Signals:  signals,
-			MfaToken: *execMfaToken,
+			Profile:     *execProfile,
+			Command:     *execCmd,
+			Args:        *execCmdArgs,
+			Keyring:     keyring,
+			Duration:    *execSessDuration,
+			Signals:     signals,
+			MfaToken:    *execMfaToken,
+			StartServer: *execServer,
 		})
 
 	case login.FullCommand():
