+++ server.go
@@ -1,12 +1,15 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
 	"io"
 	"log"
 	"net"
 	"net/http"
+	"os"
 	"os/exec"
+	"time"
 )
 
 const (
@@ -68,4 +71,48 @@ func credentialsHandler(w http.ResponseWriter, r *http.Request) {
 
 func installNetworkAlias() ([]byte, error) {
 	return exec.Command("ifconfig", "lo0", "alias", "169.254.169.254").CombinedOutput()
+}
+
+func checkServerRunning(bind string) bool {
+	_, err := net.DialTimeout("tcp", bind, time.Millisecond*10)
+	return err == nil
+}
+
+func startCredentialsServer(ui Ui, creds *VaultCredentials) error {
+	if !checkServerRunning(metadataBind) {
+		ui.Error.Println("Starting `aws-vault server` as root in the background")
+		cmd := exec.Command("sudo", "-b", os.Args[0], "server")
+		cmd.Stdin = os.Stdin
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		if err := cmd.Run(); err != nil {
+			return err
+		}
+	}
+
+	l, err := net.Listen("tcp", localServerBind)
+	if err != nil {
+		return err
+	}
+
+	log.Printf("Local instance role server running on %s", l.Addr())
+	go http.Serve(l, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		val, err := creds.Get()
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusGatewayTimeout)
+			return
+		}
+
+		json.NewEncoder(w).Encode(map[string]interface{}{
+			"Code":            "Success",
+			"LastUpdated":     time.Now().Format(awsTimeFormat),
+			"Type":            "AWS-HMAC",
+			"AccessKeyId":     val.AccessKeyID,
+			"SecretAccessKey": val.SecretAccessKey,
+			"Token":           val.SessionToken,
+			"Expiration":      creds.Expires().Format(awsTimeFormat),
+		})
+	}))
+
+	return nil
 }
